pytorch_learning/torchvision/requirements.txt
matplotlib==3.8.4
numpy==2.0.0
pandas==2.2.2
Pillow==10.3.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.65.0
